// src/payload-types.ts
export interface VercelDeployment {
  id: number;
  teamId?: string | null;
  projectId: string;
  deploymentId: string;
  deploymentDashboardUrl?: string | null;
  deploymentUrl?: string | null;
  productionDeployHookUrl?: string | null;
  projectDashboardUrl?: string | null;
  projectName: string;
  repositoryUrl?: string | null;
  updatedAt: string;
  createdAt: string;
}